moved vertical line into own class in annotations folder
